requirements.txt
django
python-dotenv
os
psycopg2
django-cors-headers
drf-yasg
djangorestframework
djangorestframework-simplejwt
django-filter